Moved paths and shim for require.config to a central location
