+++ Makefile
@@ -108,8 +108,8 @@ test: require_virtualenv_in_dev
 	PYTHONPATH=.. nosetests -a '!slow'
 
 js-test:
-	cd server/static/; \
-	python -c 'import webbrowser; webbrowser.open("http://127.0.0.1:8000/js/js_tests/test.html")'; \
+	cd server/; \
+	python -c 'import webbrowser; webbrowser.open("http://127.0.0.1:8000/static/js/js_tests/test.html")'; \
 	python -mSimpleHTTPServer 8000
 
 clean:
